requirements.txt
langchain==0.3.27
langchain-openai==0.3.35
langchain-community==0.3.31
langchain-core==0.3.83
langchain-text-splitters==0.3.11
openai==2.17.0
faiss-cpu==1.11.0
pandas==2.3.2
numpy==1.26.4
python-dotenv==1.2.1
matplotlib==3.10.3
plotly==6.6.0
streamlit==1.55.0
